requirements_dev.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2==2.9.6
